do not allow arbitratily nested synergies
as we know for sure layout of all the possible synergies we know that
the most complicated synergy requirement is a combination of two
requirements. that's where separation between "primitive" and "combined"
requirements come in. "combined" is simply two "primitive" ones together.

also introduced new term - "equipable". it unites both guns and
(active/passive) items